Refactor lookup and add reserved resolver
