+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod client;
 pub mod codec;
 pub mod database;
 pub mod error;
+pub mod lookup;
 pub mod resolver;
 pub mod rr;
 pub mod server;
